Adjust the number of defined arguments to match Winston 3+
Winston 3.0+ uses the [length of the `log` function](https://github.com/winstonjs/winston/blob/664b44ba4d8656880c1d09253eabc55636ab97ef/lib/winston/logger.js#L336) to determine whether the transport is legacy or not. In Winston 3+, the existence of `arg2` and `arg3` in the function definition results in Winston 3+'s core treating the transport as a legacy transport (which the function is _not_ set up for).

The solution is to remove the arguments from the function definition and rely on the `arguments` array to access them in legacy mode. See [this comment](https://github.com/wavded/winston-mail/issues/48#issuecomment-433471452) for more.

**NOTE:** This change assumes that the function passes the _other_ check that Winston 3+ makes to determine whethr or not to run in legacy mode:

>They inherit from `winston.Transport` in  < 3.x.x which is NOT a stream.